git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/gsoc_server_listing@18073 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/other/ares/ares_process.c
@@ -263,7 +263,7 @@ static void advance_tcp_send_queue(ares_channel channel, int whichserver,
       sendreq = server->qhead;
       if ((size_t)num_bytes >= sendreq->len)
        {
-         num_bytes -= sendreq->len;
+         num_bytes -= (ssize_t)sendreq->len;
          server->qhead = sendreq->next;
          if (server->qhead == NULL)
            {
